Routing fixes
If anyone's reading this, or for future reference, routes in razor pages are kind of weird.
Normally, it uses the file structure to determine routes. If you start a cshtml page with something like @page "{id}", it will count the id part as an addition to what the route would have been otherwise. Thus, I had to change it to @page "/products/{id}" to get the desired behavior. Starting with a / sets the route to the root of the domain, so I had to add the products, then the id.